stockmarket/requirements.txt
djongo
requests
django-cors-headers
djangorestframework
rest_framework_simplejwt
oauth2_provider